+++ src/nanasqlite/__init__.py
@@ -34,7 +34,7 @@ from .exceptions import (
     NanaSQLiteClosedError,
 )
 
-__version__ = "1.2.0dev1"
+__version__ = "1.2.0a1"
 __author__ = "Disnana"
 __all__ = [
     "NanaSQLite",